silence warnings

+++ tests/test_edge_set.cpp
@@ -43,21 +43,21 @@ public:
 TEST_F(EdgeSetTest, operatorNew)
 {
     EdgeSet* es = new EdgeSet();
-    EXPECT_EQ(es->nVertices(), 0);
-    EXPECT_EQ(es->nEdges(), 0);
+    EXPECT_EQ(es->nVertices(), size_t(0));
+    EXPECT_EQ(es->nEdges(), size_t(0));
 }
 
 TEST_F(EdgeSetTest, operatorEq)
 {
     EdgeSet es2 = es;
-    EXPECT_EQ(es2.nVertices(), 0);
+    EXPECT_EQ(es2.nVertices(), size_t(0));
 }
 
 TEST_F(EdgeSetTest, assignment)
 {
     EdgeSet es2;
     es2.assign(es);
-    EXPECT_EQ(es2.nVertices(), 0);
+    EXPECT_EQ(es2.nVertices(), size_t(0));
 }
 
 TEST_F(EdgeSetTest, addRemove)
@@ -68,11 +68,11 @@ TEST_F(EdgeSetTest, addRemove)
     auto v0  = es.addVertex(p0);
     auto v1  = es.addVertex(p1);
     auto v2  = es.addVertex(p2);
-    EXPECT_EQ(es.nVertices(), 3);
+    EXPECT_EQ(es.nVertices(), size_t(3));
 
     auto e0 = es.insertEdge(v0,v1);
     es.insertEdge(v1,v2);
-    EXPECT_EQ(es.nEdges(), 2);
+    EXPECT_EQ(es.nEdges(), size_t(2));
 
     size_t nv(0);
     for (auto v : es.vertices())
@@ -80,7 +80,7 @@ TEST_F(EdgeSetTest, addRemove)
         if (v.isValid())
             nv++;
     }
-    EXPECT_EQ(nv, 3);
+    EXPECT_EQ(nv, size_t(3));
 
     size_t ne(0);
     for (auto e : es.edges())
@@ -88,7 +88,7 @@ TEST_F(EdgeSetTest, addRemove)
         if (e.isValid())
             ne++;
     }
-    EXPECT_EQ(ne, 2);
+    EXPECT_EQ(ne, size_t(2));
 
     size_t nh(0);
     for (auto h : es.halfedges())
@@ -96,7 +96,7 @@ TEST_F(EdgeSetTest, addRemove)
         if (h.isValid())
             nh++;
     }
-    EXPECT_EQ(nh, 4);
+    EXPECT_EQ(nh, size_t(4));
 
     size_t nvv = 0;
     for (auto vv : es.vertices(v1))
@@ -104,7 +104,7 @@ TEST_F(EdgeSetTest, addRemove)
         if (vv.isValid())
             nvv++;
     }
-    EXPECT_EQ(nvv, 2);
+    EXPECT_EQ(nvv, size_t(2));
 
     size_t nvh = 0;
     for (auto vh : es.halfedges(v1))
@@ -112,15 +112,15 @@ TEST_F(EdgeSetTest, addRemove)
         if (vh.isValid())
             nvh++;
     }
-    EXPECT_EQ(nvh, 2);
+    EXPECT_EQ(nvh, size_t(2));
 
     es.deleteEdge(es.edge(e0));
     EXPECT_TRUE(es.isDeleted(e0));
     es.garbageCollection();
-    EXPECT_EQ(es.nEdges(), 1);
+    EXPECT_EQ(es.nEdges(), size_t(1));
 
     es.clear();
-    EXPECT_EQ(es.nVertices(), 0);
+    EXPECT_EQ(es.nVertices(), size_t(0));
 }
