+++ tests/test_geometry_object.cpp
@@ -40,15 +40,15 @@ TEST(GeometryObjectTest, properties)
     // explicit add
     auto midx = ps.addObjectProperty<int>("m:idx");
     midx[0] = 0;
-    EXPECT_EQ(ps.objectProperties().size(), 1);
+    EXPECT_EQ(ps.objectProperties().size(), size_t(1));
     ps.removeObjectProperty(midx);
-    EXPECT_EQ(ps.objectProperties().size(), 0);
+    EXPECT_EQ(ps.objectProperties().size(), size_t(0));
 
     // implicit add
     midx = ps.objectProperty<int>("m:idx2");
-    EXPECT_EQ(ps.objectProperties().size(), 1);
+    EXPECT_EQ(ps.objectProperties().size(), size_t(1));
     ps.removeObjectProperty(midx);
-    EXPECT_EQ(ps.objectProperties().size(), 0);
+    EXPECT_EQ(ps.objectProperties().size(), size_t(0));
 }
